Contimue to fix small issues with Redux refactor
