+++ notes/src/components/NoteCard.js
@@ -1,6 +1,7 @@
 import React from "react";
+import { Link } from "react-router-dom";
 import { connect } from "react-redux";
-import { deleteNote } from "../../actions/noteActions";
+import { deleteNote } from "../actions/noteActions";
 
 import styled from "styled-components";
 import { Card, Button } from "react-bootstrap";
@@ -46,7 +47,12 @@ const NoteCard = ({ note }) => {
   return (
     <NoteContainer>
       <Card.Body>
-        <NoteTitle>{note.title}</NoteTitle>
+        <Link
+          to={`note/${note.id}`}
+          style={{ textDecoration: "none", color: "black" }}
+        >
+          <NoteTitle>{note.title}</NoteTitle>
+        </Link>
         <hr />
         <TextContainer>{note.textBody}</TextContainer>
         <ButtonContainer>
